Back up in-progress comments file to *.bak
This change does the dumbest thing possible: make a backup of the review
file with a .bak extension before attempting to push it back up to
Github.

This addresses the main concern expressed in #10, which is that under
the current behavior if the API call to Github fails, the temp
file *still gets deleted*, and you can lose a lot of work.